pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kittycode"
version = "1.0.0"
description = "Local-first AI coding CLI with safe execution, structured memory, and model routing."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
  { name = "Krish" }
]
keywords = ["cli", "ai", "agent", "developer-tools", "terminal"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Build Tools",
  "Topic :: Utilities"
]
dependencies = [
  "typer",
  "rich",
  "pyyaml",
  "numpy",
  "python-dotenv"
]

[project.scripts]
kitty = "kittycode.cli.app:app"

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "build>=1.2.0",
  "twine>=5.0.0"
]
vector = [
  "faiss-cpu>=1.7.0",
  "sentence-transformers>=2.2.0",
]
gemini = [
  "google-genai>=0.4.0",
]
all = [
  "faiss-cpu>=1.7.0",
  "sentence-transformers>=2.2.0",
  "google-genai>=0.4.0",
]

[tool.setuptools.packages.find]
exclude = ["tests", "tests.*"]
